src/backhand/requirements.txt
flask==2.3.2
werkzeug==3.0.1
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
pycryptodome==3.20.0
